Social-Media-App-Project-ORM-SQLALCHEMY/requirements.txt
fastapi[all]
psycopg2
SQLAlchemy==1.4.54
passlib
bcrypt==4.3.0
python-jose[cryptography]
alembic==1.6.5